api/requirements.txt
flask
flask-cors
flask-jwt-extended
oracledb
python-dotenv